requirements.txt
Flask==0.10.1
Flask-Assets==0.8
Flask-SQLAlchemy==1.0
Flask-Script==0.6.6
GitHub-Flask==0.3.4
Jinja2==2.7.1
MarkupSafe==0.18
PyYAML==3.10
SQLAlchemy==0.8.4
WTForms==1.0.5
Werkzeug==0.9.4
blinker==1.3
flask-peewee==0.6.5
httplib2==0.8
itsdangerous==0.23
oauth2==1.5.211
peewee==2.1.6
requests==2.1.0
webassets==0.8
wsgiref==0.1.2
wtf-peewee==0.2.2
